# docs/openapi.json
{
  "components": {
    "schemas": {
      "CompressedRLE-Input": {
        "description": "Compressed or numeric RLE mask type.\n\nAttributes:\n    size: Mask size.\n    counts: Mask RLE encoding as compressed bytes/string or a numeric array.",
        "properties": {
          "counts": {
            "anyOf": [
              {
                "format": "binary",
                "type": "string"
              },
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              }
            ],
            "title": "Counts"
          },
          "size": {
            "items": {
              "type": "integer"
            },
            "title": "Size",
            "type": "array"
          }
        },
        "required": [
          "size",
          "counts"
        ],
        "title": "CompressedRLE",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TrackingBoxPrompt": {
        "description": "Box prompt for a tracking keyframe.",
        "properties": {
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          },
          "x": {
            "title": "X",
            "type": "integer"
          },
          "y": {
            "title": "Y",
            "type": "integer"
          }
        },
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "TrackingBoxPrompt",
        "type": "object"
      },
      "TrackingInterval": {
        "description": "Optional propagation window relative to the provided video frames.",
        "properties": {
          "direction": {
            "default": "forward",
            "enum": [
              "forward",
              "backward"
            ],
            "title": "Direction",
            "type": "string"
          },
          "endFrame": {
            "title": "Endframe",
            "type": "integer"
          },
          "startFrame": {
            "title": "Startframe",
            "type": "integer"
          }
        },
        "required": [
          "startFrame",
          "endFrame"
        ],
        "title": "TrackingInterval",
        "type": "object"
      },
      "TrackingKeyframeV1": {
        "description": "A keyframe with nested prompts, matching the Pixano frontend contract.\n\nFrame indices are 0-based **relative to the submitted media window**.",
        "properties": {
          "frameIndex": {
            "title": "Frameindex",
            "type": "integer"
          },
          "prompts": {
            "$ref": "#/components/schemas/TrackingPrompts"
          }
        },
        "required": [
          "frameIndex",
          "prompts"
        ],
        "title": "TrackingKeyframeV1",
        "type": "object"
      },
...
      "TrackingPointPrompt": {
        "description": "Point prompt for a tracking keyframe.",
        "properties": {
          "label": {
            "enum": [
              0,
              1
            ],
            "title": "Label",
            "type": "integer"
          },
          "x": {
            "title": "X",
            "type": "integer"
          },
          "y": {
            "title": "Y",
            "type": "integer"
          }
        },
        "required": [
          "x",
          "y",
          "label"
        ],
        "title": "TrackingPointPrompt",
        "type": "object"
      },
      "TrackingPrompts": {
        "description": "Prompt payload for a single keyframe (points/box XOR mask).",
        "properties": {
          "box": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrackingBoxPrompt"
              },
              {
                "type": "null"
              }
            ]
          },
          "mask": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompressedRLE-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "points": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TrackingPointPrompt"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          }
        },
        "title": "TrackingPrompts",
        "type": "object"
      },
      "TrackingRequestV1": {
        "description": "Video-tracking request with nested keyframe prompts.\n\nMedia is passed by value (URL/base64/frame list/path within an allowed media root);\ndataset-reference resolution belongs in the caller (the Pixano backend).",
        "properties": {
          "frameIndexes": {
            "items": {
              "type": "integer"
            },
            "title": "Frameindexes",
            "type": "array"
          },
          "interval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrackingInterval"
              },
              {
                "type": "null"
              }
            ]
          },
          "keyframes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TrackingKeyframeV1"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyframes"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "objectsIds": {
            "items": {
              "type": "integer"
            },
            "title": "Objectsids",
            "type": "array"
          },
          "propagate": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Propagate"
          },
          "video": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "format": "path",
                      "type": "string"
                    },
                    {
                      "format": "binary",
                      "type": "string"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "format": "path",
                "type": "string"
              },
              {
                "format": "binary",
                "type": "string"
              }
            ],
            "title": "Video"
          }
        },
        "required": [
          "model",
          "video",
          "objectsIds",
          "frameIndexes"
        ],
        "title": "TrackingRequestV1",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Pixano Inference API powered by Ray Serve",
    "title": "Pixano Inference (Ray)",
    "version": "0.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/inference/tracking/jobs": {
      "post": {
        "operationId": "submit_tracking_job_v1_inference_tracking_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackingRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Tracking Job V1 Inference Tracking Jobs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Tracking Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/inference/tracking/jobs/binary": {
      "post": {
        "operationId": "submit_tracking_job_binary_v1_inference_tracking_jobs_binary_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Tracking Job Binary V1 Inference Tracking Jobs Binary Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Tracking Job Binary",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/v1/jobs/{job_id}": {
      "delete": {
        "operationId": "cancel_job_v1_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Job V1 Jobs  Job Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Job V1 Jobs  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}